+++ app/components/design_system/nhsuk/code_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Nhsuk
+    class CodeComponent < DesignSystem::Generic::CodeComponent
+    end
+  end
+end