+++ lib/design_system/govuk.rb
@@ -2,8 +2,6 @@
 
 # This is the GOV.UK adapter for the design system
 
-require_relative 'govuk/builders/code'
-require_relative 'govuk/builders/fixed_elements'
 require_relative 'govuk/builders/pagination_renderer'
 
 require_relative 'govuk/form_builder'