+++ test/helpers/design_system_helper_test.rb
@@ -17,10 +17,22 @@ class DesignSystemHelperTest < ActionView::TestCase
   end
 
   test 'ds_fixed_elements returns correct instance' do
-    brand = 'govuk'
-    controller.stubs(brand:)
-    assert_equal @registry.builder(brand, 'fixed_elements', self).brand, ds_fixed_elements.brand
-    assert_equal @registry.builder(brand, 'FixedElements', self).brand, ds_fixed_elements.brand
+    controller.stubs(brand: 'govuk')
+
+    instance = ds_fixed_elements
+
+    assert_instance_of DesignSystem::Components::FixedElements, instance
+    assert_equal 'govuk', instance.brand
+  end
+
+  test 'ds_fixed_elements without a block can be configured and rendered' do
+    controller.stubs(brand: 'govuk')
+
+    instance = ds_fixed_elements
+    instance.main_heading 'Headings'
+    @output_buffer = instance.render
+
+    assert_select('h1.govuk-heading-xl', text: 'Headings')
   end
 
   test 'ds_fixed_elements responds to block' do